Adding the Custom Action Remember: we're not going to support this export action for all of our entities. And to make this error clearer, open AdminController - our base controller - and create a public function exportAction() that simply throws a new RuntimeException: 'Action for exporting an entity is not defined'.
If we configure everything correctly, and implement this method for all entities that need it, we should never see this error. But... just in case.

Now, to the real work. To add an export for genus, we have two options. First, in AdminController, we could create a public function exportGenusAction(). Remember, whenever EasyAdminBundle calls any of our actions - even custom actions - it always looks for that specially named method: export<EntityName>Action(). Or, we can be a bit more organized, and create a custom controller for each entity. That's what we've done already. So, in GenusController, add public function exportAction().