To save time, we've already done most of the work for the CSV export.If you downloaded the starting code, in the Service directory, you should have a CsvExporter class:
85 lines  src/AppBundle/Service/CsvExporter.php
... lines 1 - 2
namespace AppBundle\Service;
use AppBundle\Entity\Genus;
use Doctrine\Common\Collections\ArrayCollection;
use Doctrine\ORM\QueryBuilder;
use Symfony\Component\HttpFoundation\StreamedResponse;
class CsvExporter {
  public function getResponseFromQueryBuilder(QueryBuilder {dollar_sign}queryBuilder, {dollar_sign}columns, {dollar_sign}filename){
    {dollar_sign}entities = new ArrayCollection({dollar_sign}queryBuilder->getQuery()->getResult());
    {dollar_sign}response = new StreamedResponse();
    if (is_string({dollar_sign}columns)) {
      {dollar_sign}columns = {dollar_sign}this->getColumnsForEntity({dollar_sign}columns);
    }
    {dollar_sign}response->setCallback(function () use ({dollar_sign}entities, {dollar_sign}columns) {
      {dollar_sign}handle = fopen('php://output', 'w+');
      // Add header
      fputcsv({dollar_sign}handle, array_keys({dollar_sign}columns));
      while ({dollar_sign}entity = {dollar_sign}entities->current()) {
        {dollar_sign}values = [];
        foreach ({dollar_sign}columns as {dollar_sign}column => {dollar_sign}callback) {
          {dollar_sign}value = {dollar_sign}callback;
          if (is_callable({dollar_sign}callback)) {
            {dollar_sign}value = {dollar_sign}callback({dollar_sign}entity);
          }
          {dollar_sign}values[] = {dollar_sign}value;
        }
        fputcsv({dollar_sign}handle, {dollar_sign}values);
        {dollar_sign}entities->next();
      }
      fclose({dollar_sign}handle);
    });
    {dollar_sign}response->headers->set('Content-Type', 'text/csv; charset=utf-8');
    {dollar_sign}response->headers->set('Content-Disposition', 'attachment; filename= .  . ');
    return {dollar_sign}response;
  }
  private function getColumnsForEntity({dollar_sign}class){
... lines 56 - 82
  }
}
Basically, we pass it a QueryBuilder, an array of column information, or the entity's class name - which is mapped to an array of column info thanks to this special function, and the filename we want. Then, it creates the CSV and returns it as a StreamedResponse. So all we need to do is call this method and return it from our controller!

I'll paste a little bit of code in the action to get us started.

When we created the export link, we kept the existing query parameters. That means we should have a sortDirection parameter... which is a nice way of making the export order match the list order.

To create the query builder, we can actually use a protected function on the base class called createListQueryBuilder(). Pass this the entity class, either Genus::class or {dollar_sign}this->entity['class']... in case you want to make this method reusable across multiple entities. Next, pass the sort direction and then the sort field: {dollar_sign}this->request->query->get('sortField'). Finally, pass in the dql_filter option: {dollar_sign}this->entity['list']['dql_filter']. This is kind of cool. We're using the entity configuration array - which is always full of goodies - to actually read the list key and the dql_filter key below it. If we have a DQL filter on this entity, the CSV export will know about it!

Ok, finally, we're ready to use the CsvExporter class. Because I'm using the new Symfony 3.3 service configuration, the CsvExporter is already registered as a private service:

 32 lines  app/config/services.yml
... lines 1 - 5
services:
  # default configuration for services in *this* file
  _defaults:
    autowire: true
    autoconfigure: true
    public: false
  AppBundle\:
    resource: '../../src/AppBundle/*'
    exclude: '../../src/AppBundle/{Entity,Repository,Tests}'
  AppBundle\Controller\:
    resource: '../../src/AppBundle/Controller'
    public: true
    tags: ['controller.service_arguments']
... lines 21 - 32