Using DI in a Fake Action The Symfony 3.3 way of accessing a service from a controller is as an argument to the action. But... remember: this is not a real action. I mean, it's not called by the normal core, Symfony controller system. Nope, it's called by EasyAdminBundle... and none of the normal controller argument tricks work. You can't type-hint the Request or any services.
Because of this, we're going to use classic dependency injection. We can do this because this controller - well any controller if you're using the Symfony 3.3 configuration - is registered as a service. Add a __construct() function and type-hint the CsvExporter class. I'll press Alt+Enter to create a property and set it. Back down below, just return {dollar_sign}this->csvExporter->getResponseFromQueryBuilder() and pass it the , Genus::class, and genuses.csv - the filename.

Deep breath... refresh! It downloaded! Ha! In my terminal. I'll:

cat ~/Downloads/genuses.csv
There it is!

There's just one last problem: on the list page... we still have those weird export links on each row. That's technically fine... but it's super confusing. The only reason we added this export action was so that it would be a valid action to call.

Unfortunately, this also gave us those links!