Adding Form Groups Ok, back to adding groups. First, move id and slug to the end of the form. Then, on top, add a new group: type: group, css_class: 'col-sm-6', label: 'Basic Information' You can picture what this is doing: adding a div with col-sm-6, putting a header inside of it, and then printing any fields below that, but in the div.
And that's huge! Because thanks to the col-sm-6 CSS class, we can really start organizing how things look.

Move funFact and isPublished a bit further down. Then, after subFamily, add a section labeled Optional.

Yep, you can totally mix-and-match groups and sections.

At this point, funFact and isPublished will still be in the group, but they'll also be in a section within that group.