Since genusScientists is pretty big, let's put that in their own group with css_class: col-sm-6 and label: Studied by... Finally, at the bottom, add one more group. I'll use the expanded format this time: css_class: col-sm-6 and label: Identification. And yep, groups can have icon and help keys.
Phew! While I did this, I added some line breaks just so that this all looks a bit more clear: here's one group, here's a second group, the last group is at the bottom.

But what does it actually look like? Let's find out! Refresh!

Oh, this feels good. The 'Basic Information' group is on the left with the 'Optional section' at the bottom. The other two groups float to the right. Now, sometimes, you might want to force the Identification group to go onto its own line. Basically, you want to add a CSS clear after the first two groups.

To do that, on the group, add a special CSS class, called new-row. And now it floats to the next line. So, groups are a really, really neat way to control how things are rendered. It adds some nice markup, and we can add whatever classes we need. So, there's not much you can't do.