With about six lines of code, we got a fully functional admin area. It must be our birthday. But now... we need to learn how to take control and really customize!
And for most things... it's... um... easy... because EasyAdminBundle let's us control almost anything via configuration. Back on its README, scroll up and click to view the full docs.

The bundle has great docs, so we'll mostly cover the basics and then dive into the really tough stuff. Let's start with design.

Right now, in the upper left corner, it says, 'EasyAdmin'... which is probably not what I want. Like most stuff, this can be changed in the config. Add a site_name key set to AquaNote. Actually, to be fancy, add a bit of HTML in this.