When we submit a form, obviously, EasyAdminBundle is taking care of everything: handling validation, saving and adding a flash message. That's the best thing ever! Until... I need to hook into that process... then suddenly, it's the worst thing ever! What if I need to do some custom processing right before an entity is created or updated?
There are 2 main ways to hook into EasyAdminBundle...and I want you to know both. Open the User entity. It has an updatedAt field:

 281 lines  src/AppBundle/Entity/User.php
... lines 1 - 16
class User implements UserInterface {
... lines 19 - 82
  /**
   * @ORM\Column(name='updated_at', type='datetime', nullable=true)
   */
  private {dollar_sign}updatedAt;
... lines 87 - 279
}
To set this, we could use Doctrine lifecycle callbacks or a Doctrine event subscriber.

But, I want to see if we can set this instead, by hooking into EasyAdminBundle. In other words, when the user submits the form for an update, we need to run some code.

The protected AdminController Methods

The first way to do this is by adding some code to the controller. Check this out: open up the base AdminController from the bundle and search for protected function. Woh... There are a ton of methods that we can override, beyond just the actions. Like, createNewEntity(), perPersistEntity() and preUpdateEntity().

If we override preUpdateEntity() in our controller, that will be called right before any entity is updated. There are a few other cool things that you can override too.

Per-Entity Override Methods

Ok, easy! Just add preUpdateEntity() to our AdminController, right? Yep... but we can do better! If we override preUpdateEntity(), it will be called whenever any entity is updated. But we really only want it to be called for the User entity.

Once again, EasyAdminBundle has our back. Inside the base controller, search for preUpdate. Check this out: right before saving, it calls some executeDynamicMethod function and passes it preUpdate, a weird <EntityName> string, then Entity.

Actually, the bundle does this type of thing all over the place. Like above, when it calls createEditForm(). Whenever you see this, it means that bundle will first look for an entity-specific version of the method - like preUpdateUserEntity() - and call it. If that doesn't exist, it will call the normal preUpdateEntity().

This is huge: it means that each entity class can have its own set of hook methods in our AdminController!