One Controller per Entity And now that I've told you that... we're going to do something completely different. Instead of having one controller - AdminController - full of entity-specific hook methods like preUpdateUserEntity or createGenusEditForm - I prefer to create a custom controller class for each entity.
Try this: in the EasyAdmin directory, copy AdminController and rename it to UserController. Then, remove the function. Use the 'Code'->'Generate menu' - or Command+N on a Mac - to override the preUpdateEntity() method. And don't forget to update your class name to UserController. We're going to configure things so that this UserController is used only for the User admin section. And that means we can safely assume that the  argument will always be a User object. And that makes life easy: {dollar_sign}entity->setUpdatedAt(new \DateTime())