But how does EasyAdminBundle know to use this controller only for the… User entity? That happens in config.yml. Down at the bottom, under User, add controller: AppBundle\Controller\EasyAdmin\UserController
And just like that! We have one controller that's used for just our User.

Try it out! Let's go find a user... how about ID 20. Right now, its updateAt is null. Edit it... make some changes... and save! Go back to show and... we got it!

This little trick unlocks a lot of hook points. But if you look at AdminController, it's a little messy. Because, changePublishedStatusAction() is only meant to be used for the Genus class:
class AdminController extends BaseAdminController {
  public function changePublishedStatusAction(){
    {dollar_sign}id = {dollar_sign}this->request->query->get('id');
    {dollar_sign}entity = {dollar_sign}this->em->getRepository('AppBundle:Genus')->find({dollar_sign}id);
... lines 13 - 24
  }
}
But technically, this controller is being used by all entities, except User.