There's one other major way to hook into things with EasyAdminBundle. and it's my favorite! Go back to the base AdminController and search for event. You'll see a lot in here! Whenever EasyAdminBundle does, well, pretty much anything... it dispatches an event: PRE_UPDATE, POST_UPDATE, POST_EDIT, PRE_SHOW, POST_SHOW... yes we get the idea already!
And this means that we can use standard Symfony event subscribers to totally kick EasyAdminBundle's butt!

Creating an Event Subscriber

Create a new Event directory... though, this could live anywhere. Then, how about, EasyAdminSubscriber. Event subscribers always implement EventSubscriberInterface. I'll go to the 'Code'->'Generate' menu - or Command+N on a Mac - and choose 'Implement Methods' to add the one required method: getSubscribedEvents(). EasyAdminBundle dispatches a lot of events... but fortunately, they all live as constants on a helper class called EasyAdminEvents. We want to use PRE_UPDATE. Set that to execute a new method onPreUpdate that we will create in a minute.

But first, I'll hold Command and click into that class. Dude, this is cool: this puts all of the possible hook points right in front of us. There are a few different categories: most events are either for customizing the actions and views or for hooking into the entity saving process.

That difference is important, because our subscriber method will be passed slightly different information based on which event it's listening to.

Back in our subscriber, we need to create onPreUpdate(). That's easy, but it's Friday and I'm so lazy. So I'll hit the Alt+Enter shortcut and choose 'Create Method'. Thank you PhpStorm Symfony plugin!

Notice that it added a GenericEvent argument. In EasyAdminBundle, every event passes you this same object... just with different data. So, you kind of need to dump it to see what you have access to.

Since we're using Symfony 3.3 and the new service configuration, my event subscriber will automatically be loaded as a service and tagged as an event subscriber:

 32 lines  app/config/services.yml
... lines 1 - 5
services:
  # default configuration for services in *this* file
  _defaults:
    autowire: true
    autoconfigure: true
    public: false
  AppBundle\:
    resource: '../../src/AppBundle/*'
    exclude: '../../src/AppBundle/{Entity,Repository,Tests}'
  AppBundle\Controller\:
    resource: '../../src/AppBundle/Controller'
    public: true
    tags: ['controller.service_arguments']
... lines 21 - 32
If that just blew your mind, check out our Symfony 3.3 series!

This means we can just... try it! Edit a user and submit. Bam!