Since we already took care of setting the updatedAt in the controller, let's do something different. The User class also has a lastUpdatedBy field, which should be a User object:
 281 lines  src/AppBundle/Entity/User.php
... lines 1 - 16
class User implements UserInterface {
... lines 19 - 87
  /**
   * @ORM\OneToOne(targetEntity='User')
   * @ORM\JoinColumn(name='last_updated_by_id', referencedColumnName='id', nullable=true)
   */
  private {dollar_sign}lastUpdatedBy;
... lines 93 - 279
}
Let's set that here.

That means we need to get the currently-logged-in User object. To get that from inside a service, we need to use another service. At the top, add a constructor. Then, type-hint the first argument with TokenStorageInterface. Watch out: there are two of them... and oof, it's impossible to know which is which. Choose either of them for now. Then, name the argument and hit Alt+Enter to create and set a new property.

Back on top, make sure you choose the TokenStorageInterface from Security\Core\Authentication. In our method, fetch the user with {dollar_sign}user = {dollar_sign}this->tokenStorage->getToken()->getUser(). And if the User is not an instanceof our User class, that means the user isn't actually logged in. In that case, set {dollar_sign}user = null. Then, {dollar_sign}entity->setLastUpdatedBy({dollar_sign}user).

Woohoo! Thanks to the new auto-wiring stuff in Symfony 3.3, we don't need to configure anything in services.yml. Yep, with some help from the type-hint, Symfony already knows what to pass to our {dollar_sign}tokenStorage argument.

So go back, refresh and... no errors! It's always creepy when things work on the first try. Go to the show page for the User id 20. Last updated by is set!

Next, we're going to hook into the bundle further and learn how to completely disable actions based on security permissions.