Overriding the AdminController To do that, we need to sub-class the core AdminController. Create a new directory in Controller called EasyAdmin. Then inside, a new PHP class called AdminController. To make this extend the normal AdminController, add a use statement for it: use AdminController as BaseAdminController. Extend that: BaseAdminController. Next, create that action method: changePublishedStatusAction(). Notice the config key is just changePublishedStatus - EasyAdminBundle automatically expects that Action suffix.
And now that we're in a controller method... we're comfortable! I mean, we could write a killer action in our sleep. But... there's a gotcha. This method is not, exactly, like a traditional controller. That's because it's not called by Symfony's routing system... it's called directly by EasyAdminBundle, which is trying to 'fake' things.

In practice, this means one important thing: we cannot add a Request argument. Actually, all of the normal controller argument tricks will not work.. because this isn't really a real controller.

Fetching the Request & the Entity

Instead, the base AdminController has a few surprises for us: protected properties with handy things like the entity manager, the request and some EasyAdmin configuration.

Let's use this! Get the id query parameter via {dollar_sign}this->request->query->get('id'). Then, fetch the object with {dollar_sign}entity = {dollar_sign}this->em->getRepository(Genus::class)->find({dollar_sign}id). Now things are easier. Change the published status to whatever it is not currently. Then, {dollar_sign}this->em->flush(). Set a fancy flash message that says whether the genus was just published or unpublished. And finally, at the bottom, I want to redirect back to the show page. Let's go steal that code from GenusController. The one difference of course is that {dollar_sign}request needs to be {dollar_sign}this->request

This is because we haven't told Symfony to use our AdminController yet: it's still using the one from the bundle.